fixup! feat(ffi): expose `Client.sync()`
Rename to sync_v2 and create sub-module
