+++ bindings/matrix-sdk-ffi/CHANGELOG.md
@@ -41,10 +41,10 @@ All notable changes to this project will be documented in this file.
 
 ### Features
 
-- Expose sync v2 API through FFI via `Client.sync()` and
-  `Client.sync_once()`, enabling mobile clients to sync without
-  requiring Sliding Sync support on the homeserver. `Client.sync()`
-  accepts a `SyncListener` callback that receives a `SyncResponse`
+- Expose sync v2 API through FFI via `Client.sync_v2()` and
+  `Client.sync_once_v2()`, enabling mobile clients to sync without
+  requiring Sliding Sync support on the homeserver. `Client.sync_v2()`
+  accepts a `SyncListenerV2` callback that receives a `SyncResponseV2`
   after each successful sync.
   ([#6359](https://github.com/matrix-org/matrix-rust-sdk/pull/6359))
 - Expose `Room.send_state_event_raw()` for sending arbitrary state events
