+++ bindings/matrix-sdk-ffi/src/lib.rs
@@ -28,6 +28,7 @@ mod session_verification;
 mod spaces;
 mod store;
 mod sync_service;
+mod sync_v2;
 mod task_handle;
 mod timeline;
 mod utd;
